Deferred shape (#317)
* Implement legate::hash and change some of the caches to unordered_map

* Use hash maps wherever applicable

* Remove unused parameter

* Fix for clang-tidy errors

* Address review comments

* Make the use of custom hashers explicit in definitions

* Address review comments and turn some of the hashers to std::hash

* Minor clean-up

* Rename all uses of Shape to tuple<uint64_t>, as Shape will change

* Map region and field managers by index spaces instead of domains

* Initial deferred shape implementation

* Fix for clang-tidy

* Python binding for legate::Shape

* Fix the unit test

* Rename Shape::dim to ndim and fix the doc build

* Fix clang-tidy issues

* Upate match credit correctly for field managers made from unbound stores

* More clang-tidy fixes

* Docstring for Shape

* Docstring for Shape in Python as well

* Address review comments

* Shouldn't have removed inline keywords

* Check constant-ness of symbolic expressions via a proper predicate

* Should have grounded default constructed shapes to 0D

* Make clang-tidy happy

* Minor optimization for the equality check for shapes

* Eliminate unnecessary blocking in validation for alignment constraints

* Factor out the iterable-to-tuple conversion to a function

* Move conversions to and from domain and domain point to a detail header

* More documentation for Shape

* Fix the clang-tidy error

* Post merge clean-up

* Update docstring per Jeremiah's suggestion

* Field match credit must be at least 1 (which somehow got changed by a merge)

* Address review comments

* Remove an unnecessary memory allocation from issue_fill

* Update src/core/data/detail/scalar.inl


* Update src/core/data/detail/scalar.h


* Make clang-tidy happy

---------

+++ src/core/partitioning/detail/partitioner.cc
@@ -30,11 +30,11 @@ namespace legate::detail {
 
 class LaunchDomainResolver {
  private:
-  static constexpr int32_t UNSET = -1;
+  static constexpr uint32_t UNSET = -1U;
 
  public:
   void record_launch_domain(const Domain& launch_domain);
-  void record_unbound_store(int32_t unbound_dim);
+  void record_unbound_store(uint32_t unbound_dim);
   void set_must_be_sequential(bool must_be_sequential) { must_be_sequential_ = must_be_sequential; }
 
   [[nodiscard]] Domain resolve_launch_domain() const;
@@ -42,7 +42,7 @@ class LaunchDomainResolver {
  private:
   bool must_be_sequential_{};
   bool must_be_1d_{};
-  int32_t unbound_dim_{UNSET};
+  uint32_t unbound_dim_{UNSET};
   std::set<Domain> launch_domains_{};
   std::set<int64_t> launch_volumes_{};
 };
@@ -60,7 +60,7 @@ void LaunchDomainResolver::record_launch_domain(const Domain& launch_domain)
   }
 }
 
-void LaunchDomainResolver::record_unbound_store(int32_t unbound_dim)
+void LaunchDomainResolver::record_unbound_store(uint32_t unbound_dim)
 {
   if (unbound_dim_ != UNSET && unbound_dim_ != unbound_dim) {
     must_be_sequential_ = true;
@@ -89,7 +89,7 @@ Domain LaunchDomainResolver::resolve_launch_domain() const
     assert(launch_domains_.size() == 1);
   }
   auto& launch_domain = *launch_domains_.begin();
-  if (unbound_dim_ != UNSET && launch_domain.dim != unbound_dim_) {
+  if (unbound_dim_ != UNSET && launch_domain.dim != static_cast<int32_t>(unbound_dim_)) {
     return {};
   }
   return launch_domain;
